Add cancel_active_payments checkout utils and move is_fully_paid tests to correct file
